+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v1.1.0
 * Googly eyes on relics
+* Googly eyes on events
 * Googly eyes on character select screen
 * Fix: take card angle into account for eye positions
 * Eye location for mod: Disciple